monitoring/files/alert_handler/requirements.txt
slackclient
python-json-logger
boto3
